+++ passwordvault_api/passwordvault_domain/Repositories/IRepository.cs
@@ -2,7 +2,7 @@ namespace passwordvault_domain.Repositories;
 
 public interface IRepository<T>
 {
-    Task<int> Create(T entity);
+    Task<long> Create(T entity);
     Task<T> Update(T entity);
-    Task<int> Delete(int id);
+    Task<long> Delete(long id);
 }
